Minor (but important) fix to RecombineMeasurements
Measurement's operations are pointer based, which would be a big change to update at this point.  Instead, just be conscious of when operations are affecting the source measurement while producing artifacts.  This is a rare operation at the moment but likely more common in the future, so just be aware.