Major overhaul!
Okay, I found how to use the midpoint to get a consistently average 2-bit reduction - and now I have a scheme for how to exploit that in always getting a reduction.  Currently I've tested that you CAN get the data down to the target, but I'm working through how to perform the data back to where it was initially.