Began work on the passage structure
It can totally shrink arbitrary data!  It has a fail mode, but I believe that we can actually encode whatever is necessary to prevent that as a part of the composition structure.

I initially had the names of things mixed up - a movement comes after a passage, as a movement is a separate component from a composition entirely (both have a performance function)